show_results/requirements.txt
Flask
Flask-PyMongo
mysql-connector-python